requirements.txt
uvicorn==0.30.6
openai==1.44.1
flask==3.0.3
requests==2.32.3
pydantic==2.9.1
python-dotenv==1.0.1
fastapi==0.114.2